Move variant-related CatalogPromotionRule contant to Core
