+++ src/Sylius/Behat/Context/Setup/CatalogPromotionContext.php
@@ -18,11 +18,10 @@ use Doctrine\ORM\EntityManagerInterface;
 use Sylius\Behat\Service\SharedStorageInterface;
 use Sylius\Bundle\CoreBundle\Fixture\Factory\ExampleFactoryInterface;
 use Sylius\Component\Core\Model\CatalogPromotionInterface;
+use Sylius\Component\Core\Model\CatalogPromotionRuleInterface;
 use Sylius\Component\Core\Model\ChannelInterface;
 use Sylius\Component\Core\Model\ProductVariantInterface;
 use Sylius\Component\Promotion\Model\CatalogPromotionActionInterface;
-use Sylius\Component\Promotion\Model\CatalogPromotionRule;
-use Sylius\Component\Promotion\Model\CatalogPromotionRuleInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
 
 final class CatalogPromotionContext implements Context
@@ -92,7 +91,7 @@ final class CatalogPromotionContext implements Context
     {
         /** @var CatalogPromotionRuleInterface $catalogPromotionRule */
         $catalogPromotionRule = $this->catalogPromotionRuleFactory->createNew();
-        $catalogPromotionRule->setType(CatalogPromotionRule::TYPE_FOR_VARIANTS);
+        $catalogPromotionRule->setType(CatalogPromotionRuleInterface::TYPE_FOR_VARIANTS);
         $catalogPromotionRule->setConfiguration([$variant->getCode()]);
 
         $catalogPromotion->addRule($catalogPromotionRule);
