+++ src/Sylius/Bundle/CoreBundle/Resources/config/app/sylius/sylius_promotion.yml
@@ -6,6 +6,9 @@ sylius_promotion:
         catalog_promotion:
             classes:
                 model: Sylius\Component\Core\Model\CatalogPromotion
+        catalog_promotion_rule:
+            classes:
+                model: Sylius\Component\Core\Model\CatalogPromotionRule
         promotion_subject:
             classes:
                 model: "%sylius.model.order.class%"
