+++ src/Sylius/Bundle/CoreBundle/Resources/config/doctrine/model/CatalogPromotionRule.orm.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Paweł Jędrzejewski
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping">
+
+    <mapped-superclass name="Sylius\Component\Core\Model\CatalogPromotionRule" table="sylius_catalog_promotion_rule" />
+
+</doctrine-mapping>
